+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Fixed
 - Renamed `set_options` in frontend to `formatting_options` to match backend
+- Fixed double scrollbars on stats page by only using padding in `<main>`
 
 ### Changed 
 - Increased padding for leaderboard counts and widened leaderboards on the stats page